+++ src/main/java/ru/wkn/object3d/model/picture/object3d/object3dunit/Voxel.java
@@ -6,5 +6,5 @@ public interface Voxel {
 
     Pixel getPixel();
 
-    int getZ();
+    double getZ();
 }
